Add IProductRepository interface to DIP_Factory namespace
Introduce the `IProductRepository` interface to define a contract
for retrieving a list of `Product` objects, adhering to the
Dependency Inversion Principle (DIP). Added `using` directives
for `System` and `System.Collections.Generic` to enable the use
of standard .NET types and collections. Encapsulated the code
within the new `DIP_Factory` namespace.